Added page unit test, and refactored PhnInput validator.
